+++ src/Routers.tsx
@@ -1,4 +1,4 @@
-import { Login, JobSearch, JobDetail, NotFound } from '@/pages';
+import { Login, JobSearch, JobDetail, NotFound, OAuthRedirectKaKao } from '@/pages';
 import { Routes, Route } from 'react-router-dom';
 
 const Routers = () => {
@@ -7,6 +7,7 @@ const Routers = () => {
 			<Route path="/" element={<JobSearch />} />
 			<Route path="/jobs/:jobId" element={<JobDetail />} />
 			<Route path="/login" element={<Login />} />
+			<Route path="/oauth/callback/kakao" element={<OAuthRedirectKaKao />} />
 			<Route path="*" element={<NotFound />} />
 		</Routes>
 	);
