+++ src/pages/index.tsx
@@ -1,4 +1,5 @@
 export { default as Login } from './Login';
+export { default as OAuthRedirectKaKao } from './OAuthRedirectKaKao';
 export { default as JobSearch } from './JobSearch';
 export { default as JobDetail } from './JobDetail';
 export { default as NotFound } from './NotFound';
